Ensure that Figjam.adapter is set to Figjam::Rails::Application before the figjam Railtie is loaded
Figjam should be able to be used in the initializer of a Rails engine. When testing the engine using a tool like [Combustion](https://github.com/pat/combustion), we see that the Railtie to load Figjam is added immediately after the definition of `Figjam::Rails::Application`.

This can (and sometimes does) mean that the `before_configuration` block of that Railtie is immediately executed. At this point, the `Figjam.adapter` is still `nil`, so it is lazily created as a `Figjam::Application`. The Railtie then calls `Figjam.load` on it, and it blows up as `default_path` throws a `NotImplementedError`.

The solution is to set the adapter to a `Figjam::Rails::Application` as soon as possible, and before the Railtie is created.

This PR makes that 2 line swap change, and adds Combustion to the specs. If the two lines of logic are not swapped, you can see how the initializing process fails.